适配anan-boot4.0.0，升级springboot3.1、springcloud2022及相关生态组件

+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/pay/dto/PayCreateDTO.java
@@ -2,7 +2,9 @@ package top.fosin.anan.platform.modules.pay.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
-
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -10,9 +12,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import top.fosin.anan.core.util.DateTimeUtil;
 import top.fosin.anan.data.valid.group.Create;
 
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 import java.util.Date;
         
 /**
@@ -55,19 +55,19 @@ public class PayCreateDTO {
 
     @NotNull(message = "应收金额" + "{jakarta.validation.constraints.NotNull.message}", groups = Create.class)
     @Schema(description = "应收金额")
-    private Double totalMoney;
+    private BigDecimal totalMoney;
 
     @NotNull(message = "支付金额" + "{jakarta.validation.constraints.NotNull.message}", groups = Create.class)
     @Schema(description = "支付金额")
-    private Double payMoney;
+    private BigDecimal payMoney;
 
     @NotNull(message = "优惠金额" + "{jakarta.validation.constraints.NotNull.message}", groups = Create.class)
     @Schema(description = "优惠金额")
-    private Double discountMonery;
+    private BigDecimal discountMonery;
 
     @NotNull(message = "待收金额" + "{jakarta.validation.constraints.NotNull.message}", groups = Create.class)
     @Schema(description = "待收金额")
-    private Double uncollectMoney;
+    private BigDecimal uncollectMoney;
 
     @Schema(description = "付款时间")
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
