+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/pay/dto/PayOrderCreateDTO.java
@@ -2,7 +2,9 @@ package top.fosin.anan.platform.modules.pay.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
-
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -10,9 +12,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import top.fosin.anan.core.util.DateTimeUtil;
 import top.fosin.anan.data.valid.group.Create;
 
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -45,7 +45,7 @@ public class PayOrderCreateDTO {
 
     @NotNull(message = "版本金额" + "{jakarta.validation.constraints.NotNull.message}", groups = Create.class)
     @Schema(description = "版本金额")
-    private Double money;
+    private BigDecimal money;
 
     @Schema(description = "订单时间")
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
