+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/version/dto/VersionDTO.java
@@ -1,9 +1,7 @@
 package top.fosin.anan.platform.modules.version.dto;
 
 
-
 import io.swagger.v3.oas.annotations.media.Schema;
-
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -11,6 +9,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import top.fosin.anan.core.util.DateTimeUtil;
 import top.fosin.anan.data.entity.Id;
 
+import java.math.BigDecimal;
 import java.util.Date;
 /**
  * 系统版本表(anan_version)DTO
@@ -31,7 +30,7 @@ public class VersionDTO extends Id<Long> {
     private Byte type;
 
     @Schema(description = "版本价格", example = "Double")
-    private Double price;
+    private BigDecimal price;
 
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
     @Schema(description = "活动开始时间", example = DateTimeUtil.DATETIME_PATTERN)
