+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/modules/version/dto/VersionUpdateDTO.java
@@ -2,7 +2,10 @@ package top.fosin.anan.platform.modules.version.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
-
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -11,10 +14,7 @@ import top.fosin.anan.core.util.DateTimeUtil;
 import top.fosin.anan.data.entity.Id;
 import top.fosin.anan.data.valid.group.Update;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 import java.util.Date;
 /**
  * 系统版本表(anan_version)更新DTO
@@ -40,7 +40,7 @@ public class VersionUpdateDTO extends Id<Long> {
 
     @NotNull(message = "版本价格" + "{jakarta.validation.constraints.NotNull.message}", groups = Update.class)
     @Schema(description = "版本价格")
-    private Double price;
+    private BigDecimal price;
 
     @Schema(description = "活动开始时间")
     @DateTimeFormat(pattern = DateTimeUtil.DATETIME_PATTERN)
